MDA/Hercules support
If the current video mode was 80x25 mono (which is a pretty good
indicator for an MDA or Hercular card), use that mode for WorDOS
too, but with different atributes and everything shifted to the
right (so the screen is centered properly).

The chosen attributes are not particularily beautiful; I'd have
preferred to draw correctly guessed letters in bright and
underlined instead of reverse video, but I wanted to stay
compatible with true monochrome screens that don't have the
intensity bit (e.g. LCDs).